Implement 100 and 1000 iterations tests.
Also tried to implement 10^8 iterations test. But it fails. It seems
that the 10^8 iterations test vectors are doing some other operation
besides plain encryption--something to do with fiddling with the
encryption key.

I will take a closer look at the code that generates the test vectors,
and update this test code to match.